// src/types/ITicket.ts
export interface ITicket {
  id: string;
  price: number;
  carrier: string;
  carrierLogoPath: string;
  segments: [
    {
      origin: string;
      destination: string;
      date: string;
      stops: string[];
      duration: number;
    },
    {
      origin: string;
      destination: string;
      date: string;
      stops: string[];
      duration: number;
    }
  ];
}